refactor: ARCH-03 add domain errors, use domain.ErrInsufficientStock
